Optimize JSON parsing by migrating legacy nodes in database
- Add MIGRATION_22_23 to AppDatabase to migrate legacy 'thinking' nodes to 'reasoning'
- Remove runtime `migrateLegacyNodesJson` check from ConversationRepository
- Clean up unused imports
